# Performance Improvements When Parametrizing Systems