@jchodera 2019-05-21
> While this is definitely an improvement, I would recommend perhaps having a rethink of the Topology classes generator design pattern - while it should be memory efficient, in the cases I've looked at here there can be a significant compute overhead associated with it. I think a compromise between a fully computed Topology object, and a when needed generator design is definitely possible, but will likely require a rewrite of the Topology classes.

Would be great if you could capture this insight/suggestion in a separate issue!