@j-wags 2019-05-21
Excellent -- I'm deep in some 0.3 spec stuff, but will review this later today!