@SimonBoothroyd 2019-05-22
> Would be great if you could capture this insight/suggestion in a separate issue!

Sure! I added this onto the discussion in issue #253